// src/types/enums/EErrorText.ts
export enum EErrorText {
  minLength2 = 'Минимальная длина: 2 символа',
  maxLength15 = 'Максимальная длина: 15 символов',
  requiredField = 'Поле обязательно для заполнения',
  passwordMatch = 'Введенные пароли не совпадают',
  passwordSpace = 'Недопустимый формат! Ваш пароль содержит пробел',
  passwordFormat = 'Недопустимый формат! Пароль должен быть не короче 8 символов, иметь хотя бы одну заглавную и строчную буквы, одну цифру и один спец. символ',
  textFormat = 'Недопустимый формат! Пожалуйста, используйте буквы',
  invalidAge = 'Недопустимая дата рождения',
  emailFormat = 'Недопустимый формат!',
  postalCode = 'Индекс содержит 6 цифр'
}